requirements.txt
django
djangorestframework
environs
supabase
python-dotenv
psycopg2-binary==2.9.11
dj-database-url
pyotp
cryptography
